# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "qwen code API server",
    "description": "Building a blazing fast REST API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/command/mcp/list": {
      "post": {
        "tags": [
          "command",
          "mcp"
        ],
        "description": "调用mcp list命令获取MCP服务器列表",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sessionId": {
                    "default": "",
                    "type": "string"
                  },
                  "args": {
                    "default": "",
                    "type": "string"
                  }
                },
                "required": [
                  "sessionId",
                  "args"
                ],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "type": {
                      "type": "string"
                    },
                    "messageType": {
                      "type": "string"
                    },
                    "content": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "type": {
                      "type": "string"
                    },
                    "messageType": {
                      "type": "string"
                    },
                    "content": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/command/mcp": {
      "post": {
        "tags": [
          "command",
          "mcp"
        ],
        "description": "调用mcp命令获取MCP服务器列表",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sessionId": {
                    "default": "",
                    "type": "string"
                  },
                  "args": {
                    "default": "",
                    "type": "string"
                  }
                },
                "required": [
                  "sessionId",
                  "args"
                ],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "type": {
                      "type": "string"
                    },
                    "messageType": {
                      "type": "string"
                    },
                    "content": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "type": {
                      "type": "string"
                    },
                    "messageType": {
                      "type": "string"
                    },
                    "content": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}